separate testing greater into another file
