+++ .github/workflows/run-pytest-push.yml
@@ -1,13 +1,11 @@
 # This is a basic workflow to help you get started with Actions
 
-name: Run Pytest
+name: Run Pytest (Push)
 
 # Controls when the workflow will run
 on:
   # Triggers the workflow on push or pull request events but only for the master branch
   push:
-    branches: [ master ]
-  pull_request:
     branches: [ master ]
 
   # Allows you to run this workflow manually from the Actions tab
